mesh mesh intersection based algorithms exposed via occt booleans
